Web/requirements.txt
click==8.1.3
Flask==2.2.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.8.6
Werkzeug==2.2.1
zipp==3.8.1
